backend/requirements.txt
# general packages
Django==5.2.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
redis==6.4.0
celery==5.5.3
gunicorn==23.0.0
whitenoise==6.11.0

# db/dev packages
psycopg2-binary==2.9.10

# imagery packages
numpy==2.1.2
sentinelhub==3.11.1
matplotlib==3.9.4
rasterio==1.4.3
scipy==1.14.1

